+++ pages/src/components/Checkbox/Checkbox.js
@@ -7,10 +7,46 @@ var Checkbox = Backbone.View.extend({
         var view = this;
         $.get('src/components/Checkbox/checkbox.html', function(tpl) {
             view.$el.html(Mustache.render(tpl));
+			view.initBizUI();
+			hljs.initHighlighting.called = false;
+			hljs.initHighlighting();
         });
     },
 
+	initBizUI: function() {
+        var checkboxGroup = $(':checkbox');
+        checkboxGroup.bizCheckbox({
+            theme: 'green'
+        });
+        $('button').bizButton({
+            theme: 'blue-gray'
+        });
+        $('#b1').click(function() {
+            $('#apple').bizCheckbox('check');
+        });
+        $('#b2').click(function() {
+            $('#banana').bizCheckbox('check');
+        });
+        $('#b3').click(function() {
+            $('#cherry').bizCheckbox('check');
+        });
+        $('#b4').click(function() {
+            checkboxGroup.bizCheckbox('uncheck');
+        });
+        $('#b5').click(function() {
+            checkboxGroup.bizCheckbox('disable');
+        });
+        $('#b6').click(function() {
+            checkboxGroup.bizCheckbox('enable');
+        });
+        $('#b7').click(function() {
+            alert(checkboxGroup.bizCheckbox('val'));
+        });
+	},
+
     destroy: function() {
+        $('button').bizButton('destroy');
+        $(':checkbox').bizCheckbox('destroy');
         this.$el.empty();
     }
 });